requirements.txt
camelot-py==0.7.2
chardet==3.0.4
Click==7.0
et-xmlfile==1.0.1
jdcal==1.4
numpy==1.16.2
opencv-python==4.0.0.21
openpyxl==2.6.2
pandas==0.24.2
pdfminer.six==20181108
pycryptodome==3.8.0
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
sortedcontainers==2.1.0
